chore: separate different jobs to elements, use files instead of object links
Relates one-zero-eight/printers#57
